+++ Modules/FieldItems/Database/Migrations/2023_03_14_094023_add_company_id_to_item_category_table.php
@@ -28,6 +28,10 @@ return new class extends Migration
      */
     public function down()
     {
+        if (! Schema::hasTable('item_category')) {
+            return;
+        }
+        
         Schema::table('item_category', function (Blueprint $table) {
                 $table->dropForeign(['company_id']);
                 $table->dropColumn('company_id');
