+++ Modules/Payroll/Database/Migrations/2023_01_01_071222_payroll_cycle_insert_users_table.php
@@ -48,6 +48,10 @@ return new class extends Migration
      */
     public function down()
     {
+        if (! Schema::hasTable('salary_slips')) {
+            return;
+        }
+        
         Schema::table('salary_slips', function (Blueprint $table) {
             $table->dropForeign('salary_slips_currency_id_foreign');
             $table->dropColumn('currency_id');
